Add installation status API endpoint
The new endpoint returns the complete status of all non-deleted
installations managed by the provisioner. This can be used by other
cloud services to easily determine the amount of installation
processing happening at any moment so that they can make intelligent
decisions on when to request additional installation updates.

+++ model/installation.go
@@ -57,7 +57,7 @@ type Installation struct {
 
 // InstallationsCount represents the number of installations
 type InstallationsCount struct {
-	Count int
+	Count int64
 }
 
 // InstallationFilter describes the parameters used to constrain a set of installations.
@@ -191,7 +191,7 @@ func InstallationsFromReader(reader io.Reader) ([]*Installation, error) {
 
 // InstallationsCountFromReader decodes a json-encoded installations count data from the
 // given io.Reader
-func InstallationsCountFromReader(reader io.Reader) (int, error) {
+func InstallationsCountFromReader(reader io.Reader) (int64, error) {
 	installationsCount := InstallationsCount{}
 	decoder := json.NewDecoder(reader)
 	err := decoder.Decode(&installationsCount)
